chore: extract css and define config into reusable functions
